pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karya"
version = "0.1.0"
description = "Offline autonomous agent for low-power hardware (Raspberry Pi, edge devices)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "agent", "raspberry-pi", "edge-ai", "offline", "autonomous", "iot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: No Input/Output (Daemon)",
]

dependencies = []  # zero required deps — stdlib only for core

[project.optional-dependencies]
full = [
    "pyyaml>=6.0",      # goals.yaml parsing
    "psutil>=5.9",      # better system metrics
]
gpio = [
    "RPi.GPIO>=0.7",    # Raspberry Pi GPIO (Pi only)
    "gpiozero>=2.0",    # alternative GPIO library
]
serial = [
    "pyserial>=3.5",    # serial port support
]
all = [
    "pyyaml>=6.0",
    "psutil>=5.9",
    "pyserial>=3.5",
]

[project.urls]
Homepage = "https://github.com/yourusername/karya"
Issues = "https://github.com/yourusername/karya/issues"
Changelog = "https://github.com/yourusername/karya/blob/main/CHANGELOG.md"

[project.scripts]
karya = "karya.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["karya*"]

[tool.setuptools.package-data]
karya = ["py.typed"]
